+++ library/src/main/java/com/getstream/sdk/chat/rest/interfaces/SearchMessagesCallback.java
@@ -1,0 +1,9 @@
+package com.getstream.sdk.chat.rest.interfaces;
+
+import com.getstream.sdk.chat.rest.response.SearchMessagesResponse;
+
+public interface SearchMessagesCallback {
+    void onSuccess(SearchMessagesResponse response);
+
+    void onError(String errMsg, int errCode);
+}