+++ library/src/main/java/com/getstream/sdk/chat/viewmodel/ChannelViewModel.java
@@ -688,6 +688,23 @@ public class ChannelViewModel extends AndroidViewModel implements LifecycleHandl
             messages.postValue(messagesCopy);
     }
 
+    private void checkFailedMessage(Message message){
+        if (message.getSyncStatus() != Sync.LOCAL_FAILED)
+            return;
+
+        List<Message> messagesCopy = getMessages().getValue();
+        for (int i = 0; i < messagesCopy.size(); i++) {
+            if (message.getId().equals(messagesCopy.get(i).getId())) {
+                messagesCopy.remove(message);
+                if (isThread())
+                    threadMessages.postValue(messagesCopy);
+                else
+                    messages.postValue(messagesCopy);
+                break;
+            }
+        }
+    }
+
     private void addMessage(Message message) {
         List<Message> messagesCopy = getMessages().getValue();
         messagesCopy.add(message);
@@ -945,8 +962,13 @@ public class ChannelViewModel extends AndroidViewModel implements LifecycleHandl
         if (message.getSyncStatus() == LOCAL_ONLY) {
             return;
         }
+
         // Check Error or Pending Messages
         checkErrorOrPendingMessage();
+
+        // Check Failed Message
+        checkFailedMessage(message);
+
         // stop typing
         stopTyping();
         // Check uploading file