+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Upcoming release
+
+- Added support for message search `client.searchMessages`
+- Better support for query user options
+- Minor bugfixes
+
 ## Oct 16th, 2019 - 2.3.0
 
 - Added support for `getReactions` endpoint